refractor dashboard treasury to fix inconsistent data
